pace-cv-engine/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "pace-cv-engine"
version = "0.1.0"
description = "PACE Platform CV Microservice — MediaPipe pose estimation + face masking pipeline"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.111.0",
    "uvicorn[standard]==0.30.1",
    "mediapipe==0.10.14",
    "opencv-python-headless==4.10.0.84",
    "numpy==1.26.4",
    "boto3==1.34.144",
    "aiobotocore==2.13.1",
    "pydantic==2.7.4",
    "pydantic-settings==2.3.4",
    "httpx==0.27.0",
    "python-multipart==0.0.9",
    "structlog==24.2.0",
    "sentry-sdk[fastapi]==2.7.1",
    "pgmpy==0.1.26",
    "networkx==3.3",
    "pandas==2.2.2",
    "scikit-learn==1.5.1",
]

[project.optional-dependencies]
dev = [
    "pytest==8.2.2",
    "pytest-asyncio==0.23.7",
    "pytest-cov==5.0.0",
    "httpx==0.27.0",
    "ruff==0.5.0",
    "mypy==1.10.1",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
